refactor: Move transfer-related classes to the regular transfer package(code)
