+++ application/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/request/transfercache/RaptorRequestTransferCacheKey.java
@@ -2,7 +2,6 @@ package org.opentripplanner.routing.algorithm.raptoradapter.transit.request.tran
 
 import java.util.List;
 import java.util.Objects;
-import org.opentripplanner.routing.algorithm.raptoradapter.transit.Transfer;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.street.model.StreetMode;
 import org.opentripplanner.street.search.request.BikeRequest;
@@ -11,6 +10,7 @@ import org.opentripplanner.street.search.request.StreetSearchRequest;
 import org.opentripplanner.street.search.request.WalkRequest;
 import org.opentripplanner.street.search.request.WheelchairRequest;
 import org.opentripplanner.streetadapter.StreetSearchRequestMapper;
+import org.opentripplanner.transfer.regular.model.Transfer;
 import org.opentripplanner.utils.tostring.ToStringBuilder;
 
 class RaptorRequestTransferCacheKey {
